allow multiple config sections per file
Assuming it's ok to disallow ":" in filenames, this change would allow config section names for files to have trailing characters that are not interpreted as part of the filename, but that make the key unique in terms of config parsing, e.g.:

[bumpversion:file:myfile:0]
search = some context for {current_version}

[bumpversion:file:myfile:1]
search = another context for {current_version}

In our application, we have two different sets of parse/serialize/search/replace configs that need to be run on the same files (e.g. some versions appear as "major.minor.release" while others appear as "major_minor_release", and with different surrounding search context), and this one minor change to bumpversion would make such a scenario possible.